move duotone methods into duotone module
