+++ src/plugins/api-proxy/util/duotone.js
@@ -110,3 +110,63 @@ export const getDuotoneUrls = (duotones, PHOTO_SCALER_SALT) => {
 		{}
 	);
 };
+
+/**
+ * From a provided set of signed duotone URLs, create a function that injects
+ * the full duotone URL into a group object with the key `duotoneUrl`.
+ *
+ * @param {Object} duotoneUrls map of `[duotoneRef]: url template root`
+ * @param {Object} group group object from API
+ * @return {Object} the mutated group object
+ */
+export const groupDuotoneSetter = duotoneUrls => group => {
+	const photo = group.key_photo || {};
+	const duotoneKey =
+		group.photo_gradient &&
+		duotoneRef(
+			group.photo_gradient.light_color,
+			group.photo_gradient.dark_color
+		);
+	const duotoneUrlRoot = duotoneKey && duotoneUrls[duotoneKey];
+	if (duotoneUrlRoot && photo.id) {
+		group.duotoneUrl = `${duotoneUrlRoot}/${photo.id}.jpeg`;
+	}
+	return group;
+};
+
+/**
+ * From a provided set of signed duotoneUrls, create a function that injects
+ * the full duotone URL into an query response containing objects that support
+ * duotoned images (anything containing group or event objects
+ *
+ * @param {Object} duotoneUrls map of `[duotoneRef]: url template root`
+ * @param {Object} queryResponse { ref, type: <type>, value: <API object>, error? }
+ * @return {Object} the modified queryResponse
+ */
+export const apiResponseDuotoneSetter = duotoneUrls => {
+	const setGroupDuotone = groupDuotoneSetter(duotoneUrls);
+	return queryResponse => {
+		// inject duotone URLs into any group query response
+		const { type, value, error } = queryResponse;
+		if (!value || error) {
+			return queryResponse;
+		}
+		let groups;
+		switch (type) {
+			case 'group':
+				groups = value instanceof Array ? value : [value];
+				groups.forEach(setGroupDuotone);
+				break;
+			case 'home':
+				(value.rows || [])
+					.map(({ items }) => items)
+					.forEach(items =>
+						items
+							.filter(({ type }) => type === 'group')
+							.forEach(({ group }) => setGroupDuotone(group))
+					);
+				break;
+		}
+		return queryResponse;
+	};
+};
